requirements.txt
jose
passlib
app==0.0.1
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
fastapi==0.65.2
gunicorn==20.1.0
h11==0.12.0
idna==2.10
importlib-metadata==4.6.1
nanodb==0.4.3
nanodb-driver==0.0.3
pyasn1==0.4.8
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.1
PyYAML==3.11
pyzmq==22.1.0
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
utils==1.0.1
uvicorn==0.14.0
zipp==3.5.0
